Crate sp_partner_chains_bridge

Primitives and inherent data provider for the token bridge feature of Partner Chains Toolkit.

Structs

MainChainScripts

Smart contract hashes and addresses used by the token bridge on Cardano.

TokenBridgeTransfersV1

Structure representing all token bridge transfers incoming from Cardano that are to be handled in one Partner Chain block.

Enums

BridgeDataCheckpoint

Value specifying the point in time up to which bridge transfers have been processed

BridgeTransferV1

Type containing all information needed to process a single transfer incoming from main chain, corresponding to a single UTXO on Cardano

InherentDataCreationError

Error type returned when creating TokenBridgeInherentDataProvider fails

InherentError

Error type returned by failing calls of the bridge inherent

TokenBridgeInherentDataProvider

Inherent data provider that provides data on token bridge transfers and special token transfers

Constants

INHERENT_IDENTIFIER

Inherent identifier used by the Partner Chains token bridge

Traits

TokenBridgeDataSource

Interface for data sources that can be used by TokenBridgeInherentDataProvider

TokenBridgeIDPRuntimeApi

Runtime API used by TokenBridgeInherentDataProvider